adasd

+++ controleretorno/src/main/java/br/com/doctorfeet/controleretorno/entity/Feedback.java
@@ -29,7 +29,7 @@ public class Feedback implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
 
     @Column(name = "nome")
     private String nome;
@@ -40,15 +40,15 @@ public class Feedback implements Serializable {
     public Feedback() {
     }
 
-    public Feedback(Integer id) {
+    public Feedback(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
